+++ src/components/ui-parts/OptionalUI.tsx
@@ -3,12 +3,14 @@ import { FC } from 'react';
 import {
   Control,
   ControllerRenderProps,
+  useFieldArray,
   UseFormRegister,
 } from 'react-hook-form';
 import RoyaltyTextField from '../textField/RoyaltyTextField';
 import CreatorUI from './CreatorUI';
-import { validationRules } from '../../shared/validation';
+import AddCreatorButton from '../button/AddCreatorButton';
 import { NFTFormValues } from '../../pages/NftPage';
+import { validationRules } from '../../shared/validation';
 
 export type OptionalUIProps = {
   isShow: boolean;
@@ -18,6 +20,19 @@ export type OptionalUIProps = {
 };
 
 const OptionalUI: FC<OptionalUIProps> = ({ isShow, control }) => {
+  const { fields, append, remove } = useFieldArray({
+    name: 'creators',
+    control,
+  });
+
+  const handleAddButton = () => {
+    append({
+      address: '',
+      share: 0,
+      verified: false,
+    });
+  };
+
   return isShow ? (
     <>
       <Box sx={{ mb: 4 }} />
@@ -27,8 +42,9 @@ const OptionalUI: FC<OptionalUIProps> = ({ isShow, control }) => {
         rules={validationRules.royalty}
       />
       <Box sx={{ mb: 4 }} />
-      <CreatorUI {...{ control }} />
+      <CreatorUI {...{ control, fields, remove }} />
       <Box sx={{ mb: 4 }} />
+      <AddCreatorButton callbackFunc={handleAddButton} />
     </>
   ) : null;
 };
